requirements.txt
Django==2.2.10
django-filter==2.0.0
djangorestframework==3.9.1
djangorestframework-csv==2.1.0
djangorestframework-xml==1.3.0
djangorestframework-yaml==1.0.3
django-cors-headers
six==1.11.0
